+++ src/logger.ts
@@ -1,15 +1,18 @@
 import winston, {format} from 'winston';
+import {Config} from './config';
 
 const prettyJson = format.printf(info => {
+	const timestamp = new Date().toLocaleTimeString();
+
 	if (typeof info.message === 'object') {
 		info.message = JSON.stringify(info.message, null, 4);
 	}
 
-	return `${info.level} :: ${info.message}`;
+	return `[${timestamp}] ${info.level} :: ${info.message}`;
 });
 
 export const Logger = winston.createLogger({
-	level: process.env.LOG_LEVEL ?? 'info',
+	level: Config.logLevel,
 	format: format.combine(
 		format.colorize(),
 		format.prettyPrint(),