+++ src/notification/email.ts
@@ -3,23 +3,24 @@ import Mail from 'nodemailer/lib/mailer';
 import {Config} from '../config';
 import {Logger} from '../logger';
 
+const email = Config.notifications.email;
 const subject = 'NVIDIA - BUY NOW';
 
 const transporter = nodemailer.createTransport({
 	service: 'gmail',
 	auth: {
-		user: Config.notifications.email.username,
-		pass: Config.notifications.email.password
+		user: email.username,
+		pass: email.password
 	}
 });
 
 const mailOptions: Mail.Options = {
-	from: Config.notifications.email.username,
-	to: Config.notifications.email.username,
+	from: email.username,
+	to: email.username,
 	subject
 };
 
-export default function sendEmail(text: string) {
+export function sendEmail(text: string) {
 	mailOptions.text = text;
 
 	transporter.sendMail(mailOptions, (error, info) => {