+++ src/store/includes-labels.ts
@@ -1,0 +1,10 @@
+/**
+ * Checks if DOM has any out-of-stock related text.
+ *
+ * @param domText Complete DOM of website.
+ * @param oosLabels Out-of-stock labels.
+ */
+export function includesLabels(domText: string, oosLabels: string[]): boolean {
+	const domTextLowerCase = domText.toLowerCase();
+	return oosLabels.some(label => domTextLowerCase.includes(label));
+}